tests: move host vs guest cpu features into a single test
This is the same test spread over two files. Putting it together in a
new file so they don't drift over time.
